+++ lib/screen/result_page.dart
@@ -1,5 +1,5 @@
 import 'package:color_tile/provider/block_provider.dart';
-import 'package:color_tile/provider/elapsedtime_provider.dart';
+import 'package:color_tile/provider/time_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
